+++ xbmc/games/controllers/dialogs/ControllerInstaller.cpp
