+++ xbmc/games/controllers/dialogs/ControllerInstaller.h
