+++ xbmc/games/ports/input/CMakeLists.txt
@@ -1,0 +1,11 @@
+set(SOURCES PhysicalPort.cpp
+            PortInput.cpp
+            PortManager.cpp
+)
+
+set(HEADERS PhysicalPort.h
+            PortInput.h
+            PortManager.h
+)
+
+core_add_library(games_ports_input)
