+++ xbmc/games/ports/input/PortInput.cpp
@@ -8,8 +8,8 @@
 
 #include "PortInput.h"
 
-#include "InputSink.h"
 #include "games/addons/GameClient.h"
+#include "games/controllers/input/InputSink.h"
 #include "guilib/WindowIDs.h"
 #include "input/joysticks/keymaps/KeymapHandling.h"
 #include "peripherals/devices/Peripheral.h"
