+++ xbmc/games/ports/types/CMakeLists.txt
@@ -1,0 +1,7 @@
+set(SOURCES PortNode.cpp
+)
+
+set(HEADERS PortNode.h
+)
+
+core_add_library(games_ports_types)
