+++ xbmc/games/ports/windows/CMakeLists.txt
@@ -1,0 +1,11 @@
+set(SOURCES GUIPortList.cpp
+            GUIPortWindow.cpp
+)
+
+set(HEADERS GUIPortDefines.h
+            GUIPortList.h
+            GUIPortWindow.h
+            IPortList.h
+)
+
+core_add_library(games_ports_windows)
