+++ xbmc/games/ports/windows/GUIPortDefines.h
