+++ xbmc/games/ports/windows/GUIPortList.h
@@ -8,11 +8,11 @@
 
 #pragma once
 
-#include "ControllerSelect.h"
 #include "IPortList.h"
 #include "addons/AddonEvents.h"
 #include "games/GameTypes.h"
 #include "games/controllers/ControllerTypes.h"
+#include "games/controllers/dialogs/ControllerSelect.h"
 #include "games/controllers/types/ControllerTree.h"
 
 #include <map>
