+++ xbmc/games/ports/windows/GUIPortWindow.cpp
