+++ xbmc/games/ports/windows/GUIPortWindow.h
